Improved del functionality and number input
Now the delete button can remove operators all together as opposed to before. Also if the user enters a number after getting a result, the result will be wiped from the screen automatically, if an operator is entered, the operator will be appended to the end of the result making further calculations possible.